feat: compatible with webpack4
