+++ packages/electron-plugin/bridge.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/apis/bridge.js');
