+++ packages/electron-plugin/main.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/apis/main.js');
