+++ packages/electron-plugin/test.main.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/apis/test.main.js');
