+++ packages/electron-plugin/test.render.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/apis/test.render.js');
