+++ packages/electron-runtime/render.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/index.render.js');
