+++ packages/electron-test/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/index.js');
