+++ packages/electron-test/webview.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/js/node/index.webview.js');
